+++ core/class/frigate.class.php
@@ -113,12 +113,12 @@ class frigate extends eqLogic
 
   private static function execCron($frequence)
   {
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:START CRON:/fg:----------------------------------");
-    log::add(__CLASS__, 'debug', "| Exécution du cron : {$frequence}");
+    log::add(__CLASS__, 'debug', "╔════════════════════════fg-success:START CRON:/fg:════════════════════════");
+    log::add(__CLASS__, 'debug', "║ Exécution du cron : {$frequence}");
     if (config::byKey("cron::run", 'frigate')) {
-      log::add(__CLASS__, 'debug', "| Un cron est deja en cours d'exécution, on n'exécute pas de nouveau.");
+      log::add(__CLASS__, 'debug', "║ Un cron est deja en cours d'exécution, on n'exécute pas de nouveau.");
       config::save('cron::run', 0, 'frigate');
-      log::add(__CLASS__, 'debug', "----------------------END CRON----------------------------------");
+      log::add(__CLASS__, 'debug', "╚════════════════════════END CRON════════════════════════");
       return;
     }
     config::save('cron::run', 1, 'frigate');
@@ -130,9 +130,9 @@ class frigate extends eqLogic
         $frequence === "functionality::cron5::enable" ||
         $frequence === "functionality::cron10::enable" ||
         $frequence === "functionality::cron15::enable")) {
-        log::add(__CLASS__, 'debug', "| Les crons 1, 5, 10 et 15 sont désactivés avec MQTT et ne sont pas utilisés.");
+        log::add(__CLASS__, 'debug', "║ Les crons 1, 5, 10 et 15 sont désactivés avec MQTT et ne sont pas utilisés.");
         config::save('cron::run', 0, 'frigate');
-        log::add(__CLASS__, 'debug', "----------------------END CRON----------------------------------");
+        log::add(__CLASS__, 'debug', "╚════════════════════════END CRON═══════════════════");
         return;
       }
     }
@@ -167,23 +167,24 @@ class frigate extends eqLogic
     }
 
     config::save('cron::run', 0, 'frigate');
-    log::add(__CLASS__, 'debug', "----------------------END CRON----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════END CRON═══════════════════");
   }
 
 
-  private static function isFrigateServerAvailable() {
+  private static function isFrigateServerAvailable()
+  {
     $urlFrigate = self::getUrlFrigate();
     if ($urlFrigate === false) {
       return false;
     }
-    
+
     $ch = curl_init("http://" . $urlFrigate . "/api/version");
     curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
     curl_setopt($ch, CURLOPT_TIMEOUT, 5);
     $result = curl_exec($ch);
     $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
     curl_close($ch);
-    
+
     return $httpCode >= 200 && $httpCode < 300 && !empty($result);
   }
 
@@ -394,7 +395,7 @@ class frigate extends eqLogic
   {
     $name = $this->getConfiguration('name');
     $events = frigate_events::all();
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:START REMOVE EQLOGIC:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:START REMOVE EQLOGIC:/fg:═══════════════════");
     foreach ($events as $event) {
       if ($event->getCamera() == $name) {
         $event->setIsFavorite(0);
@@ -403,7 +404,7 @@ class frigate extends eqLogic
         self::cleanDbEvent($eventId);
       }
     }
-    log::add(__CLASS__, 'debug', "----------------------END REMOVE EQLOGIC----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════END REMOVE EQLOGIC═══════════════════");
   }
 
   /*
@@ -767,12 +768,12 @@ class frigate extends eqLogic
   {
     $url = config::byKey('URL', 'frigate');
     if ($url == "") {
-      log::add(__CLASS__, 'error', "| Erreur: L'URL ne peut être vide.");
+      log::add(__CLASS__, 'error', "║ Erreur: L'URL ne peut être vide.");
       return false;
     }
     $port = config::byKey('port', 'frigate');
     if ($port == "") {
-      log::add(__CLASS__, 'error', "| Erreur: Le port ne peut être vide");
+      log::add(__CLASS__, 'error', "║ Erreur: Le port ne peut être vide");
       return false;
     }
     $urlFrigate = $url . ":" . $port;
@@ -799,7 +800,7 @@ class frigate extends eqLogic
 
   public static function publish_message(string $subTopic, string $payload)
   {
-    log::add(__CLASS__, 'debug', "| publish_message : " . self::getTopic() . "/{$subTopic} avec payload : {$payload}");
+    log::add(__CLASS__, 'debug', "║ publish_message : " . self::getTopic() . "/{$subTopic} avec payload : {$payload}");
     mqtt2::publish(self::getTopic() . "/{$subTopic}", $payload);
   }
 
@@ -826,12 +827,12 @@ class frigate extends eqLogic
     $data = curl_exec($ch);
 
     if (curl_errno($ch)) {
-      log::add(__CLASS__, 'error', "| Erreur getcURL (" . $method . "): " . curl_error($ch));
+      log::add(__CLASS__, 'error', "║ Erreur getcURL (" . $method . "): " . curl_error($ch));
       die();
     }
     curl_close($ch);
     $response = $decodeJson ? json_decode($data, true) : $data;
-    log::add(__CLASS__, 'debug', "| " . $function . " : requête " . $method . " exécutée.");
+    log::add(__CLASS__, 'debug', "║ " . $function . " : requête " . $method . " exécutée.");
     return $response;
   }
 
@@ -855,33 +856,33 @@ class frigate extends eqLogic
     $data = curl_exec($ch);
 
     if (curl_errno($ch)) {
-      log::add(__CLASS__, 'error', "| Erreur: deletecURL" . curl_error($ch));
+      log::add(__CLASS__, 'error', "║ Erreur: deletecURL" . curl_error($ch));
       die();
     }
     curl_close($ch);
     $response = json_decode($data, true);
-    log::add(__CLASS__, 'debug', "| Suppression sur le serveur Frigate : " . json_encode($response));
+    log::add(__CLASS__, 'debug', "║ Suppression sur le serveur Frigate : " . json_encode($response));
     return $response;
   }
 
   public static function getStats()
   {
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:START STATS:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:START STATS:/fg:═══════════════════");
     $urlfrigate = self::getUrlFrigate();
     $resultURL = $urlfrigate . "/api/stats";
     $stats = self::getcURL("Stats", $resultURL);
     self::majStatsCmds($stats);
-    log::add(__CLASS__, 'debug', "----------------------END STATS----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════END STATS═══════════════════");
   }
 
   public static function getPresets($camera)
   {
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:START IMPORT PRESETS:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:START IMPORT PRESETS:/fg:═══════════════════");
     $urlfrigate = self::getUrlFrigate();
     $resultURL = $urlfrigate . "/api/" . $camera . "/ptz/info";
     $presets = self::getcURL("Presets", $resultURL);
+    log::add(__CLASS__, 'debug', "╚════════════════════════END IMPORT PRESET----------------------------");
     return $presets;
-    log::add(__CLASS__, 'debug', "----------------------END IMPORT PRESET----------------------------");
   }
 
   public static function createEvent($camera, $label, $video = 1, $duration = 20, $score = 30, $subLabel = '')
@@ -894,13 +895,13 @@ class frigate extends eqLogic
     $duration = floatval($duration);
     $includeRecording = ($video == 1);
 
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:START CREATE EVENT:/fg:----------------------------------");
-    log::add(__CLASS__, 'debug', "| label : {$label}");
-    log::add(__CLASS__, 'debug', "| score : {$score}");
-    log::add(__CLASS__, 'debug', "| duration : {$duration}");
-    log::add(__CLASS__, 'debug', "| video : {$video}");
-    log::add(__CLASS__, 'debug', "| include_recording : " . ($includeRecording ? "true" : "false"));
-    log::add(__CLASS__, 'debug', "| sub_label : {$subLabel}");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:START CREATE EVENT:/fg:═══════════════════");
+    log::add(__CLASS__, 'debug', "║ label : {$label}");
+    log::add(__CLASS__, 'debug', "║ score : {$score}");
+    log::add(__CLASS__, 'debug', "║ duration : {$duration}");
+    log::add(__CLASS__, 'debug', "║ video : {$video}");
+    log::add(__CLASS__, 'debug', "║ include_recording : " . ($includeRecording ? "true" : "false"));
+    log::add(__CLASS__, 'debug', "║ sub_label : {$subLabel}");
 
     $params = [
       'source_type' => 'api',
@@ -911,7 +912,7 @@ class frigate extends eqLogic
     ];
     $response = self::postcURL("CreateEvent", $resultURL, $params);
 
-    log::add(__CLASS__, 'debug', "----------------------END CREATE EVENT----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════END CREATE EVENT═══════════════════");
     return $response;
   }
 
@@ -920,16 +921,16 @@ class frigate extends eqLogic
   // TODO : Ajouter des méthodes appelant cette méthode pour une modification de paramètres du fichier de configuration
   public static function saveConfig($config)
   {
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:START SAVE CONFIG:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:START SAVE CONFIG:/fg:═══════════════════");
     $urlfrigate = self::getUrlFrigate();
     $resultURL = $urlfrigate . "/api/config/set?{$config}";
 
-    log::add(__CLASS__, 'debug', "| url : {$resultURL}");
+    log::add(__CLASS__, 'debug', "║ url : {$resultURL}");
     $response = self::putcURL("saveConfig", $resultURL); //, $params);
 
     event::add('frigate::config', array('message' => 'api_config_update', 'type' => 'config'));
 
-    log::add(__CLASS__, 'debug', "----------------------END SAVE CONFIG----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════END SAVE CONFIG═══════════════════");
     return $response;
   }
 
@@ -1000,12 +1001,12 @@ class frigate extends eqLogic
     foreach ($filteredRecoveryEvents as $event) {
       $frigate = frigate_events::byEventId($event['id']);
 
-      log::add(__CLASS__, 'debug', "----------------------:fg-success:START EVENT:/fg:----------------------------------");
+      log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:START EVENT:/fg:═══════════════════");
 
       $infos = self::getEventinfos($mqtt, $event, false, $type);
 
       if (!$frigate) {
-        log::add(__CLASS__, 'debug', "| Events (type=" . $type . ") => " . json_encode($event));
+        log::add(__CLASS__, 'debug', "║ Events (type=" . $type . ") => " . json_encode($event));
         $box = $event['data']['box'] ?? "null";
 
         $frigate = new frigate_events();
@@ -1033,7 +1034,7 @@ class frigate extends eqLogic
         $frigate->setIsFavorite(0);
         $frigate->save();
         self::majEventsCmds($frigate);
-        log::add(__CLASS__, 'debug', "| Evénement Frigate créé et sauvegardé, event ID: " . $event['id']);
+        log::add(__CLASS__, 'debug', "║ Evénement Frigate créé et sauvegardé, event ID: " . $event['id']);
       } else {
 
         if (is_array($frigate) && !empty($frigate)) {
@@ -1074,18 +1075,18 @@ class frigate extends eqLogic
             if ($value !== null) {
               $newValue = json_encode($value);
             }
-            // log::add(__CLASS__, 'debug', "| BOX, ancienne valeur: " . $currentValue . ", nouvelle valeur: " . $newValue);
+            // log::add(__CLASS__, 'debug', "║ BOX, ancienne valeur: " . $currentValue . ", nouvelle valeur: " . $newValue);
           }
 
           if ((is_null($currentValue) || $currentValue === '' || $currentValue != $newValue) && !is_null($newValue) && $newValue !== '') {
-            log::add(__CLASS__, 'debug', "| Mise à jour du champ '$field' pour event ID: " . $event['id'] . ". ancienne valeur: " . json_encode($currentValue) . ", nouvelle valeur: " . json_encode($newValue));
+            log::add(__CLASS__, 'debug', "║ Mise à jour du champ '$field' pour event ID: " . $event['id'] . ". ancienne valeur: " . json_encode($currentValue) . ", nouvelle valeur: " . json_encode($newValue));
             $frigate->$setMethod($newValue);
             $updated = true;
             if ($field == 'Type' && $newValue != $frigate->getType()) {
               $infos = self::getEventinfos($mqtt, $event, true);
               $frigate->setSnapshot($infos["snapshot"]);
               $frigate->setClip($infos["clip"]);
-              log::add(__CLASS__, 'debug', "| Mise à jour forcé des champs snapshot et clip pour event ID: " . $event['id']);
+              log::add(__CLASS__, 'debug', "║ Mise à jour forcé des champs snapshot et clip pour event ID: " . $event['id']);
               $frigate->save();
             }
           }
@@ -1093,15 +1094,15 @@ class frigate extends eqLogic
 
         if ($updated) {
           $frigate->setData($event['data']);
-          log::add(__CLASS__, 'debug', "| Mise à jour du champ data pour event ID: " . $event['id']);
+          log::add(__CLASS__, 'debug', "║ Mise à jour du champ data pour event ID: " . $event['id']);
           $frigate->save();
           self::majEventsCmds($frigate);
-          log::add(__CLASS__, 'debug', "| Evénement Frigate mis à jour et sauvegardé, event ID: " . $event['id']);
+          log::add(__CLASS__, 'debug', "║ Evénement Frigate mis à jour et sauvegardé, event ID: " . $event['id']);
         } else {
-          log::add(__CLASS__, 'debug', "| Pas de mise à jour pour event ID: " . $event['id']);
+          log::add(__CLASS__, 'debug', "║ Pas de mise à jour pour event ID: " . $event['id']);
         }
       }
-      log::add(__CLASS__, 'debug', "----------------------END EVENT----------------------------------");
+      log::add(__CLASS__, 'debug', "╚════════════════════════END EVENT═══════════════════");
     }
   }
 
@@ -1198,7 +1199,7 @@ class frigate extends eqLogic
   {
     $filePath = $dir . '/' . $id . $suffix;
     if (!file_exists($filePath)) {
-      log::add(__CLASS__, 'debug', "| Fichier non trouvé: $filePath, téléchargement");
+      log::add(__CLASS__, 'debug', "║ Fichier non trouvé: $filePath, téléchargement");
       sleep(5); // Option à améliorer
       $img = self::saveURL($id, $isThumbnail ? null : "snapshot", $camera, $isThumbnail);
       return $img == "error" ? "null" : "/plugins/frigate/data/" . $camera . "/" . $id . $suffix;
@@ -1209,14 +1210,14 @@ class frigate extends eqLogic
   private static function processSnapshot($dir, $event, $force)
   {
     if (!file_exists($dir . '/' . $event['id'] . '_snapshot.jpg') || $force) {
-      log::add(__CLASS__, 'debug', "| Fichier snapshot non trouvé: " . $dir . '/' . $event['id'] . '_snapshot.jpg');
+      log::add(__CLASS__, 'debug', "║ Fichier snapshot non trouvé: " . $dir . '/' . $event['id'] . '_snapshot.jpg');
       if ($event['has_snapshot'] == "true") {
-        log::add(__CLASS__, 'debug', "| Has Snapshot: true, téléchargement");
+        log::add(__CLASS__, 'debug', "║ Has Snapshot: true, téléchargement");
         sleep(5); // Option à améliorer
         $snapshot = self::saveURL($event['id'], "snapshot", $event['camera']);
         return ['url' => $snapshot == "error" ? "null" : $snapshot, 'has' => $snapshot != "error"];
       }
-      log::add(__CLASS__, 'debug', "| Has Snapshot: false, téléchargement annulé");
+      log::add(__CLASS__, 'debug', "║ Has Snapshot: false, téléchargement annulé");
     }
     return ['url' => "/plugins/frigate/data/" . $event['camera'] . "/" . $event['id'] . '_snapshot.jpg', 'has' => 1];
   }
@@ -1224,12 +1225,12 @@ class frigate extends eqLogic
   private static function processClip($dir, $event, $type, $force)
   {
     if ($type != "end") {
-      log::add(__CLASS__, 'debug', "| Pas de clip, le type n'est pas 'end' " . json_encode($event));
+      log::add(__CLASS__, 'debug', "║ Pas de clip, le type n'est pas 'end' " . json_encode($event));
       return ['url' => "null", 'has' => 0];
     }
 
     if (!file_exists($dir . '/' . $event['id'] . '_clip.mp4') || $force) {
-      log::add(__CLASS__, 'debug', "| Fichier clip non trouvé: " . $dir . '/' . $event['id'] . '_clip.mp4');
+      log::add(__CLASS__, 'debug', "║ Fichier clip non trouvé: " . $dir . '/' . $event['id'] . '_clip.mp4');
       if ($event['has_clip'] == "true") {
         sleep(5); // Option à améliorer
         $clip = self::saveURL($event['id'], "clip", $event['camera']);
@@ -1237,11 +1238,11 @@ class frigate extends eqLogic
 
         $duration = self::getVideoDuration($dir . '/' . $event['id'] . '_clip.mp4');
         if ($duration !== false) {
-          log::add(__CLASS__, 'debug', "| La durée de la video est de " . gmdate("H:i:s", $duration));
+          log::add(__CLASS__, 'debug', "║ La durée de la video est de " . gmdate("H:i:s", $duration));
         }
         return ['url' => $clip, 'has' => 1];
       } else {
-        log::add(__CLASS__, 'debug', "| Has Clip: false, telechargement annulé");
+        log::add(__CLASS__, 'debug', "║ Has Clip: false, telechargement annulé");
         return ['url' => "null", 'has' => 0];
       }
     }
@@ -1249,7 +1250,7 @@ class frigate extends eqLogic
   private static function processPreview($dir, $event)
   {
     if (!file_exists($dir . '/' . $event['id'] . '_preview.gif')) {
-      log::add(__CLASS__, 'debug', "| Fichier preview non trouvé: " . $dir . '/' . $event['id'] . '_preview.gif');
+      log::add(__CLASS__, 'debug', "║ Fichier preview non trouvé: " . $dir . '/' . $event['id'] . '_preview.gif');
       $preview = self::saveURL($event['id'], "preview", $event['camera']);
       return ['url' => $preview == "error" ? "null" : $preview, 'has' => $preview != "error"];
     }
@@ -1295,18 +1296,18 @@ class frigate extends eqLogic
             // Vérifier si l'id existe dans la base de données
             $frigate = frigate_events::byEventId($id);
             if (!$frigate) {
-              log::add(__CLASS__, 'debug', "| Fichier " . $path . " non trouvé en database.");
+              log::add(__CLASS__, 'debug', "║ Fichier " . $path . " non trouvé en database.");
               if (unlink($path)) {
-                log::add(__CLASS__, 'debug', "| Suppresion reussie: " . $path);
+                log::add(__CLASS__, 'debug', "║ Suppresion reussie: " . $path);
               } else {
-                log::add(__CLASS__, 'error', "| Suppresion echouée: " . $path);
+                log::add(__CLASS__, 'error', "║ Suppresion echouée: " . $path);
               }
             }
           }
         }
       }
     } else {
-      log::add(__CLASS__, 'error', "| Dossier inexistant: " . $folder);
+      log::add(__CLASS__, 'error', "║ Dossier inexistant: " . $folder);
     }
   }
 
@@ -1318,14 +1319,14 @@ class frigate extends eqLogic
     $days = config::byKey('remove_days', 'frigate', "7");
     $recoveryDays = config::byKey('recovery_days', 'frigate', "7");
     if (!is_numeric($days) || $days <= 0) {
-      log::add(__CLASS__, 'error', "| Configuration invalide pour 'remove_days': " . $days . " Cela doit être un nombre positif.");
+      log::add(__CLASS__, 'error', "║ Configuration invalide pour 'remove_days': " . $days . " Cela doit être un nombre positif.");
       return;
     }
     if ($days < $recoveryDays) {
-      log::add(__CLASS__, 'warning', "| 'remove_days' doit être supérieur à 'recovery_days'");
+      log::add(__CLASS__, 'warning', "║ 'remove_days' doit être supérieur à 'recovery_days'");
       $days = $recoveryDays;
     }
-    log::add(__CLASS__, 'info', "| Nettoyage des fichiers datant de plus de " . $days . " jours.");
+    log::add(__CLASS__, 'info', "║ Nettoyage des fichiers datant de plus de " . $days . " jours.");
 
     $events = frigate_events::getOldestNotFavorites($days);
 
@@ -1333,18 +1334,18 @@ class frigate extends eqLogic
       foreach ($events as $event) {
         $eventId = $event->getEventId();
 
-        log::add(__CLASS__, 'info', "| Nettoyage de l'événement ID: " . $eventId);
+        log::add(__CLASS__, 'info', "║ Nettoyage de l'événement ID: " . $eventId);
 
         $result = self::cleanDbEvent($eventId);
 
         if ($result) {
-          log::add(__CLASS__, 'info', "| Événement ID: " . $eventId . " nettoyé avec succès.");
+          log::add(__CLASS__, 'info', "║ Événement ID: " . $eventId . " nettoyé avec succès.");
         } else {
-          log::add(__CLASS__, 'error', "| Échec du nettoyage de l'événement ID: " . $eventId);
+          log::add(__CLASS__, 'error', "║ Échec du nettoyage de l'événement ID: " . $eventId);
         }
       }
     } else {
-      log::add(__CLASS__, 'info', "| Aucun événement trouvé datant de plus de " . $days . " jours.");
+      log::add(__CLASS__, 'info', "║ Aucun événement trouvé datant de plus de " . $days . " jours.");
     }
   }
 
@@ -1356,14 +1357,14 @@ class frigate extends eqLogic
       foreach ($events as $event) {
         $eventId = $event->getEventId();
 
-        log::add(__CLASS__, 'info', "| Nettoyage de l'événement ID: " . $eventId . " il est de type: " . $type);
+        log::add(__CLASS__, 'info', "║ Nettoyage de l'événement ID: " . $eventId . " il est de type: " . $type);
 
         $result = self::cleanDbEvent($eventId);
 
         if ($result) {
-          log::add(__CLASS__, 'info', "| Événement ID: " . $eventId . " nettoyé avec succès.");
+          log::add(__CLASS__, 'info', "║ Événement ID: " . $eventId . " nettoyé avec succès.");
         } else {
-          log::add(__CLASS__, 'error', "| Échec du nettoyage de l'événement ID: " . $eventId);
+          log::add(__CLASS__, 'error', "║ Échec du nettoyage de l'événement ID: " . $eventId);
         }
       }
     }
@@ -1385,18 +1386,18 @@ class frigate extends eqLogic
   {
     $maxSize = config::byKey('datas_weight', 'frigate');
     $size = self::getFolderSize();
-    log::add(__CLASS__, 'debug', "| Taille du dossier : " . $size);
-    log::add(__CLASS__, 'debug', "| Taille maximale du dossier : " . $maxSize);
+    log::add(__CLASS__, 'debug', "║ Taille du dossier : " . $size);
+    log::add(__CLASS__, 'debug', "║ Taille maximale du dossier : " . $maxSize);
 
     while ($size > $maxSize) {
-      log::add(__CLASS__, 'debug', "| Le dossier est plein, nettoyage du fichier le plus ancien");
+      log::add(__CLASS__, 'debug', "║ Le dossier est plein, nettoyage du fichier le plus ancien");
       self::cleanOldestFile();
       $size = self::getFolderSize();
-      log::add(__CLASS__, 'debug', "| Nouvelle taille du dossier : " . $size);
+      log::add(__CLASS__, 'debug', "║ Nouvelle taille du dossier : " . $size);
     }
 
     if ($size <= $maxSize) {
-      log::add(__CLASS__, 'debug', "| Le dossier n'est pas plein");
+      log::add(__CLASS__, 'debug', "║ Le dossier n'est pas plein");
     }
   }
 
@@ -1437,7 +1438,7 @@ class frigate extends eqLogic
       if (
         $isFavorite == 1
       ) {
-        log::add(__CLASS__, 'debug', "| Événement " . $frigate->getEventId() . " est un favori, il ne doit pas être supprimé de la base de données.");
+        log::add(__CLASS__, 'debug', "║ Événement " . $frigate->getEventId() . " est un favori, il ne doit pas être supprimé de la base de données.");
       } else {
         // Recherche si clip et snapshot existent dans le dossier de sauvegarde
         $clip = dirname(__FILE__, 3) . "/data/" . $frigate->getCamera() . "/" . $frigate->getEventId() . "_clip.mp4";
@@ -1447,23 +1448,23 @@ class frigate extends eqLogic
 
         if (file_exists($clip)) {
           unlink($clip);
-          log::add(__CLASS__, 'debug', "| Clip MP4 supprimé pour l'événement " . $frigate->getEventId());
+          log::add(__CLASS__, 'debug', "║ Clip MP4 supprimé pour l'événement " . $frigate->getEventId());
         }
         if (file_exists($snapshot)) {
           unlink($snapshot);
-          log::add(__CLASS__, 'debug', "| Snapshot JPG supprimé pour l'événement " . $frigate->getEventId());
+          log::add(__CLASS__, 'debug', "║ Snapshot JPG supprimé pour l'événement " . $frigate->getEventId());
         }
         if (file_exists($thumbnail)) {
           unlink($thumbnail);
-          log::add(__CLASS__, 'debug', "| Miniature JPG supprimée pour l'événement " . $frigate->getEventId());
+          log::add(__CLASS__, 'debug', "║ Miniature JPG supprimée pour l'événement " . $frigate->getEventId());
         }
         if (file_exists($preview)) {
           unlink($preview);
-          log::add(__CLASS__, 'debug', "| GIF supprimé pour l'événement " . $frigate->getEventId());
+          log::add(__CLASS__, 'debug', "║ GIF supprimé pour l'événement " . $frigate->getEventId());
         }
 
         $frigate->remove();
-        log::add(__CLASS__, 'debug', "| Événement " . $frigate->getEventId() . " supprimé de la base de données.");
+        log::add(__CLASS__, 'debug', "║ Événement " . $frigate->getEventId() . " supprimé de la base de données.");
       }
       return true;
     } else {
@@ -1484,7 +1485,7 @@ class frigate extends eqLogic
     if (!empty($frigate) && isset($frigate[0])) {
       $isFavorite = $frigate[0]->getIsFavorite() ?? 0;
       if ($isFavorite == 1) {
-        log::add(__CLASS__, 'debug', "| Evènement " . $frigate[0]->getEventId() . " est un favori, il ne doit pas être supprimé de la DB.");
+        log::add(__CLASS__, 'debug', "║ Evènement " . $frigate[0]->getEventId() . " est un favori, il ne doit pas être supprimé de la DB.");
         message::add('frigate', __("L'évènement est un favori, il ne peut pas être supprimé de la DB.", __FILE__), null, null);
         return "Error 01";
       }
@@ -1558,22 +1559,22 @@ class frigate extends eqLogic
   public static function generateAllEqs()
   {
 
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:CREATION DES EQUIPEMENTS:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:CREATION DES EQUIPEMENTS:/fg:═══════════════════");
     $urlfrigate = self::getUrlFrigate();
     // récupérer le json de configuration
     $configurationArray = self::jsonFromUrl("http://" . $urlfrigate . "/api/config");
-    log::add(__CLASS__, 'debug', "| Fichier de configuration : " . json_encode($configurationArray));
+    log::add(__CLASS__, 'debug', "║ Fichier de configuration : " . json_encode($configurationArray));
 
     frigate::generateEqEvents($configurationArray);
     frigate::generateEqStats();
     frigate::generateEqCameras($configurationArray);
 
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:FIN CREATION DES EQUIPEMENTS:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════:fg-success:FIN CREATION DES EQUIPEMENTS:/fg:═══════════════════");
   }
   public static function generateEqCameras($configurationArray)
   {
 
-    log::add(__CLASS__, 'debug', "----------------------:fg-success:CREATION DES CAMERAS:/fg:----------------------------------");
+    log::add(__CLASS__, 'debug', "╔════════════════════════:fg-success:CREATION DES CAMERAS:/fg:═══════════════════");
     $urlfrigate = self::getUrlFrigate();
     $mqttCmds = isset($configurationArray['mqtt']['host']) && !empty($configurationArray['mqtt']['host']);
     $addToName = "";
@@ -1594,7 +1595,7 @@ class frigate extends eqLogic
         }
       }
       if ($exist) {
-        log::add(__CLASS__, 'debug', "| L'équipement : " . json_encode($cameraName) . " existe dans la pièce : " . jeeObject::byId($defaultRoom)->getName());
+        log::add(__CLASS__, 'debug', "║ L'équipement : " . json_encode($cameraName) . " existe dans la pièce : " . jeeObject::byId($defaultRoom)->getName());
         $addToName = " by frigate plugin";
       }
       // Recherche équipement caméra
@@ -1625,18 +1626,18 @@ class frigate extends eqLogic
         if ($defaultRoom) $frigate->setObject_id($defaultRoom);
         $frigate->setIsEnable(1);
         $frigate->setIsVisible(1);
-        log::add(__CLASS__, 'debug', "| L'équipement : " . json_encode($cameraName . $addToName) . " est créé.");
+        log::add(__CLASS__, 'debug', "║ L'équipement : " . json_encode($cameraName . $addToName) . " est créé.");
       } else {
-        log::add(__CLASS__, 'debug', "| L'équipement : " . json_encode($cameraName) . " n'est pas créé.");
+        log::add(__CLASS__, 'debug', "║ L'équipement : " . json_encode($cameraName) . " n'est pas créé.");
       }
       $frigate->setLogicalId("eqFrigateCamera_" . $cameraName);
       $frigate->save();
       // commandes identique pour toutes les caméras
-      log::add(__CLASS__, 'debug', "| Création des commandes génèrales pour : " . json_encode($cameraName));
+      log::add(__CLASS__, 'debug', "║ Création des commandes génèrales pour : " . json_encode($cameraName));
       self::createCamerasCmds($frigate->getId());
       // commandes MQTT s'il est configuré
       if ($mqttCmds) {
-        log::add(__CLASS__, 'debug', "| Création des commandes MQTT pour : " . json_encode($cameraName));
+        log::add(__CLASS__, 'debug', "║ Création des commandes MQTT pour : " . json_encode($cameraName));
         $value["detect"] = isset($cameraConfig['detect']['enabled']) ? $cameraConfig['detect']['enabled'] : $configurationArray['detect']['enabled'];
         $value["recordings"] = isset($cameraConfig['record']['enabled']) ? $cameraConfig['record']['enabled'] : $configurationArray['record']['enabled'];
         $value["snapshots"] = isset($cameraConfig['snapshots']['enabled']) ? $cameraConfig['snapshots']['enabled'] : $configurationArray['snapshots']['enabled'];
@@ -1657,7 +1658,7 @@ class frigate extends eqLogic
         }
         // commande PTZ si onvif est configuré
         if (isset($cameraConfig['onvif']['host']) && !empty($cameraConfig['onvif']['host']) && $cameraConfig['onvif']['host'] !== '0.0.0.0') {
-          log::add(__CLASS__, 'debug', "| Création des commandes PTZ pour : " . json_encode($cameraName));
+          log::add(__CLASS__, 'debug', "║ Création des commandes PTZ pour : " . json_encode($cameraName));
           self::createPTZcmds($frigate->getId());
           self::createPresetPTZcmds($frigate->getId());
           $frigate->setConfiguration("ptz", 1);
@@ -1668,7 +1669,7 @@ class frigate extends eqLogic
       $isAudioEnabledGlobally  = isset($configurationArray['audio']['enabled']) && !empty($configurationArray['audio']['enabled']);
       $isAudioEnabledForCamera = isset($cameraConfig['audio']['enabled_in_config']) && !empty($cameraConfig['audio']['enabled_in_config']);
       if ($isAudioEnabledGlobally  || $isAudioEnabledForCamera) {
-        log::add(__CLASS__, 'debug', "| Création des commandes audio pour : " . json_encode($cameraName));
+        log::add(__CLASS__, 'debug', "║ Création des commandes audio pour : " . json_encode($cameraName));
 
         $valueAudio = $isAudioEnabledForCamera ? $cameraConfig['audio']['enabled'] : $configurationArray['audio']['enabled'];
 
@@ -1682,7 +1683,7 @@ class frigate extends eqLogic
     self::getEvents(false, array(), 'end', null, 1);
     message::add('frigate', 'Mise à jour des commandes, évènements et statistiques terminé.');
 
-    log::add(__CLASS__, 'debug', "----------------------END CREATION DES CAMERAS----------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════END CREATION DES CAMERAS═══════════════════");
     return $n;
   }
 
@@ -2078,7 +2079,7 @@ class frigate extends eqLogic
     // Création des commandes jusqu'au nombre max de presets configurés
     for ($i = 0; $i < $max && $i < count($presetList); $i++) {
       $presetName = $presetList[$i];
-      log::add(__CLASS__, 'debug', "| PRESET CREE . " . $presetName); // Utiliser le nom du preset correspondant
+      log::add(__CLASS__, 'debug', "║ PRESET CREE . " . $presetName); // Utiliser le nom du preset correspondant
       // Vérifier que le nom du preset est une chaîne de caractères valide
       if (is_string($presetName) && !empty($presetName)) {
         $cmd = self::createCmd($eqlogicId, $presetName, "other", "", "action_preset_" . $i, "CAMERA_PRESET", 1, "", 0, "action");
@@ -2133,15 +2134,15 @@ class frigate extends eqLogic
       }
       // Vérifier si la liste d'actions est vide
       $cameraActionsExist = !empty($cameraAction);
-      log::add(__CLASS__, 'debug', "| EVENT :: ADD.");
+      log::add(__CLASS__, 'debug', "║ EVENT :: ADD.");
       self::eventAdd($event, $eqCamera->getId());
-      log::add(__CLASS__, 'debug', "| EVENT :: END. = "  . json_encode($event));
+      log::add(__CLASS__, 'debug', "║ EVENT :: END. = "  . json_encode($event));
     }
 
     $eqCamera->getId();
 
     if ($cameraActionsExist) {
-      log::add(__CLASS__, 'debug', "| ACTION: Vérification des actions caméra.");
+      log::add(__CLASS__, 'debug', "║ ACTION: Vérification des actions caméra.");
 
       // Vérifier si toutes les actions sont désactivées
       $allActionsDisabled = true;
@@ -2152,7 +2153,7 @@ class frigate extends eqLogic
         if ($enable) {
           // Si au moins une action est activée, on met à jour $allActionsDisabled à false
           $allActionsDisabled = false;
-          log::add(__CLASS__, 'debug', "| ACTION: Une action caméra est activée.");
+          log::add(__CLASS__, 'debug', "║ ACTION: Une action caméra est activée.");
           break;
         }
       }
@@ -2160,20 +2161,20 @@ class frigate extends eqLogic
       // Si toutes les actions sont désactivées, on met à jour $cameraActionsExist à false
       if ($allActionsDisabled) {
         $cameraActionsExist = false;
-        log::add(__CLASS__, 'debug', "| ACTION: Toutes les actions caméra sont désactivées.");
+        log::add(__CLASS__, 'debug', "║ ACTION: Toutes les actions caméra sont désactivées.");
       }
     } else {
-      log::add(__CLASS__, 'debug', "| ACTION: Aucune action configurée.");
+      log::add(__CLASS__, 'debug', "║ ACTION: Aucune action configurée.");
     }
 
     // Vérification des actions caméra existantes
     // Si la liste d'actions n'est pas vide et qu'au moins une action est activée
     if ($cameraActionsExist) {
-      log::add(__CLASS__, 'debug', "| ACTION: Exécution des actions pour la caméra (ID: " . $eqCamera->getId() . ").");
+      log::add(__CLASS__, 'debug', "║ ACTION: Exécution des actions pour la caméra (ID: " . $eqCamera->getId() . ").");
       self::executeActionNewEvent($eqCamera->getId(), $event);
     } else {
       // Sinon, on exécute les actions suivantes
-      log::add(__CLASS__, 'debug', "| ACTION: Aucune action caméra activée, exécution des actions pour l'équipement Events (ID: " . $frigate->getId() . ").");
+      log::add(__CLASS__, 'debug', "║ ACTION: Aucune action caméra activée, exécution des actions pour l'équipement Events (ID: " . $frigate->getId() . ").");
       self::executeActionNewEvent($frigate->getId(), $event);
     }
 
@@ -2382,11 +2383,18 @@ class frigate extends eqLogic
         $cmdLabelName = $action['cmdLabelName'] ?: "all";
         $cmdTypeName = $action['cmdTypeName'] ?: "end";
         $cmdZoneName = $action['cmdZoneName'] ?: "all";
+        $cmdZoneEndName = $action['cmdZoneEndName'] ?: "";
 
         // Convertir les chaînes en tableaux
-        $cmdLabels = array_map('trim', explode(',', $cmdLabelName));
+        /*  $cmdLabels = array_map('trim', explode(',', $cmdLabelName));
         $cmdZones = array_map('trim', explode(',', $cmdZoneName));
-        $eventZones = array_map('trim', explode(',', $zones));
+        $cmdZonesEnd = array_map('trim', explode(',', $cmdZoneEndName));
+        $eventZones = array_map('trim', explode(',', $zones)); */
+
+        $cmdLabels = array_map(fn($s) => self::cleanString(trim($s)), explode(',', $cmdLabelName));
+        $cmdZones = array_map(fn($s) => self::cleanString(trim($s)), explode(',', $cmdZoneName));
+        $cmdZonesEnd = array_map(fn($s) => self::cleanString(trim($s)), explode(',', $cmdZoneEndName));
+        $eventZones = array_map(fn($s) => self::cleanString(trim($s)), explode(',', $zones));
 
         // Ajouter "all" aux tableaux si nécessaire
         if (in_array("all", $cmdLabels)) $cmdLabels[] = $label;
@@ -2395,28 +2403,55 @@ class frigate extends eqLogic
         // Vérifier les trois conditions
         $labelMatch = in_array($label, $cmdLabels) || in_array("all", $cmdLabels);
         $typeMatch = ($cmdTypeName === $type);
-        $zoneMatch = count(array_intersect($cmdZones, $eventZones)) > 0;
+        // Verifier si on utilise zone end, si non utilisé gestion classique sinon verifier ordre des zones
+        if (empty($cmdZoneEndName)) {
+          $zoneMatch = count(array_intersect($cmdZones, $eventZones)) > 0;
+        } else {
+          // Récupérer les zones configurées
+          $enterZone = $cmdZones[0]; // Zone d'entrée configurée
+          $quitZone = $cmdZonesEnd[0]; // Zone de sortie configurée
+
+          // Trouver les positions des zones dans la séquence
+          $enterZonePos = array_search($enterZone, $eventZones);
+          $quitZonePos = array_search($quitZone, $eventZones);
+
+          // Vérifier que les deux zones sont présentes et dans le bon ordre
+          $zoneMatch = ($enterZonePos !== false && $quitZonePos !== false && $enterZonePos < $quitZonePos);
+          log::add("frigate-zones", 'debug', "╔════════════════════════════════════════════════════════════════╗");
+          log::add("frigate-zones", 'debug', "║ Zones de l'évènement : " . json_encode($eventZones));
+          log::add("frigate-zones", 'debug', "║ Zone d'entrée' : " . json_encode($enterZone));
+          log::add("frigate-zones", 'debug', "║ Zone de sortie : " . json_encode($quitZone));
+          if ($zoneMatch) {
+            log::add("frigate-zones", 'debug', "║ Correspondance trouvé, déclenchement des actions.");
+          } else {
+
+            log::add("frigate-zones", 'debug', "║ Les zones ne correspondent pas !");
+          }
+          log::add("frigate-zones", 'debug',  "║ Types de l'event : " . json_encode($type) );
+          log::add("frigate-zones", 'debug',  "║ Types de cmd event : " . json_encode($cmdTypeName));
+          log::add("frigate-zones", 'debug', "╚════════════════════════════════════════════════════════════════╝");
+        }
 
         if (!($labelMatch && $typeMatch && $zoneMatch)) {
-          log::add(__CLASS__, 'debug', "| ACTION: Au moins une des conditions (label, type, zone) n'est pas remplie, l'action sera ignorée.");
+          log::add(__CLASS__, 'debug', "║ ACTION: Au moins une des conditions (label, type, zone) n'est pas remplie, l'action sera ignorée.");
           continue;
         }
-        
+
         $options = $action['options'];
         $enable = $action['options']['enable'] ?? false;
         $actionForced = $action['options']['actionForced'] ?? false;
 
         if (!$enable) {
-          log::add(__CLASS__, 'debug', "| Commande(s) désactivée(s)");
+          log::add(__CLASS__, 'debug', "║ Commande(s) désactivée(s)");
           continue;
         }
 
         if (!$conditionIsActived) {
-          log::add(__CLASS__, 'debug', "| Commande(s) exécutée(s)");
+          log::add(__CLASS__, 'debug', "║ Commande(s) exécutée(s)");
         } elseif ($conditionIsActived && $actionForced) {
-          log::add(__CLASS__, 'debug', "| Commande(s) exécutée(s) car la condition est ignorée");
+          log::add(__CLASS__, 'debug', "║ Commande(s) exécutée(s) car la condition est ignorée");
         } else {
-          log::add(__CLASS__, 'info', "| " . $eqLogic->getHumanName() . ' : actions non exécutées car ' . $conditionIf . ' est vrai.');
+          log::add(__CLASS__, 'info', "║ " . $eqLogic->getHumanName() . ' : actions non exécutées car ' . $conditionIf . ' est vrai.');
           continue;
         }
 
@@ -2428,12 +2463,12 @@ class frigate extends eqLogic
 
         // Vérifie si le temps de début de l'événement est inférieur ou égal à trois heures avant le temps actuel
         if ($event->getStartTime() <= time() - 10800) {
-          log::add(__CLASS__, 'debug', "| ACTION: Événement trop ancien (plus de 3 heures), il sera ignoré.");
+          log::add(__CLASS__, 'debug', "║ ACTION: Événement trop ancien (plus de 3 heures), il sera ignoré.");
           continue;
         }
 
         if (!($labelMatch && $typeMatch && $zoneMatch)) {
-          log::add(__CLASS__, 'debug', "| ACTION: Au moins une des conditions (label, type, zone) n'est pas remplie, l'action sera ignorée.");
+          log::add(__CLASS__, 'debug', "║ ACTION: Au moins une des conditions (label, type, zone) n'est pas remplie, l'action sera ignorée.");
           continue;
         }
 
@@ -2441,16 +2476,16 @@ class frigate extends eqLogic
         $optionsJson = json_encode($action['options']);
         if (strpos($optionsJson, '#clip#') !== false || strpos($optionsJson, '#clip_path#') !== false) {
           if ($hasClip == 1) {
-            log::add(__CLASS__, 'debug', "| ACTION CLIP : " . $optionsJson);
+            log::add(__CLASS__, 'debug', "║ ACTION CLIP : " . $optionsJson);
             scenarioExpression::createAndExec('action', $cmd, $options);
           }
         } elseif (strpos($optionsJson, '#snapshot#') !== false || strpos($optionsJson, '#snapshot_path#') !== false) {
           if ($hasSnapshot == 1) {
-            log::add(__CLASS__, 'debug', "| ACTION SNAPSHOT : " . $optionsJson);
+            log::add(__CLASS__, 'debug', "║ ACTION SNAPSHOT : " . $optionsJson);
             scenarioExpression::createAndExec('action', $cmd, $options);
           }
         } else {
-          log::add(__CLASS__, 'debug', "| ACTION : " . $optionsJson);
+          log::add(__CLASS__, 'debug', "║ ACTION : " . $optionsJson);
           scenarioExpression::createAndExec('action', $cmd, $options);
         }
       }
@@ -2485,7 +2520,7 @@ class frigate extends eqLogic
       $lien = $file;
       $path = "/data/" . $camera . "/latest.jpg";
     } elseif ($mode == 3) {
-     // $file = $file . '?timestamp=1&bbox=1';
+      // $file = $file . '?timestamp=1&bbox=1';
       $lien = urldecode($file);
       $path = "/data/snapshots/" . $eventId . "_snapshot.jpg";
     } elseif ($mode == 4) {
@@ -2497,9 +2532,9 @@ class frigate extends eqLogic
       $cmd = 'ffmpeg -rtsp_transport tcp -loglevel fatal -i "' . $file . '" -c:v copy -bsf:a aac_adtstoasc -y -t 10 -movflags faststart ' . $newpath;
       exec($cmd, $output, $return_var);
       $result = "/plugins/frigate" . $path;
-      log::add(__CLASS__, 'debug', "| Commande exécutée : " . $cmd);
-      log::add(__CLASS__, 'debug', "| Sortie : " . implode("\n", $output));
-      log::add(__CLASS__, 'debug', "| Code de retour : " . $return_var);
+      log::add(__CLASS__, 'debug', "║ Commande exécutée : " . $cmd);
+      log::add(__CLASS__, 'debug', "║ Sortie : " . implode("\n", $output));
+      log::add(__CLASS__, 'debug', "║ Code de retour : " . $return_var);
       return $result;
     }
 
@@ -2514,7 +2549,7 @@ class frigate extends eqLogic
     // Vérifier si le répertoire existe, sinon le créer
     if (!is_dir($destinationDir)) {
       if (!mkdir($destinationDir, 0755, true)) {
-        log::add(__CLASS__, 'debug', "| Échec de la création du répertoire.");
+        log::add(__CLASS__, 'debug', "║ Échec de la création du répertoire.");
         return $result;
       }
     }
@@ -2533,23 +2568,31 @@ class frigate extends eqLogic
       $file = file_put_contents(dirname(__FILE__, 3) . $path, $content);
       if ($file !== false) {
         $result = "/plugins/frigate" . $path;
-        log::add(__CLASS__, 'debug', "| Le fichier a été enregistré : " . $lien);
+        log::add(__CLASS__, 'debug', "║ Le fichier a été enregistré : " . $lien);
       } else {
-        log::add(__CLASS__, 'debug', "| Échec de l'enregistrement du fichier : " . $lien);
+        log::add(__CLASS__, 'debug', "║ Échec de l'enregistrement du fichier : " . $lien);
         $result = "error";
       }
     } else {
-      log::add(__CLASS__, 'debug', "| Le fichier n'existe pas ou une erreur s'est produite.");
+      log::add(__CLASS__, 'debug', "║ Le fichier n'existe pas ou une erreur s'est produite.");
       $result = "error";
     }
 
     return $result;
   }
 
+  private static function cleanString($string)
+  {
+    // Supprimer les accents
+    $string = iconv('UTF-8', 'ASCII//TRANSLIT', $string);
+    // Mettre en minuscule
+    return mb_strtolower($string);
+  }
+
   public static function createSnapshot($eqLogic)
   {
-    log::add(__CLASS__, 'debug', "---------------------------------------------------");
-    log::add(__CLASS__, 'debug', "| Créer snapshot");
+    log::add(__CLASS__, 'debug', "╔════════════════════════════════════════════════");
+    log::add(__CLASS__, 'debug', "║ Créer snapshot");
     $camera = $eqLogic->getConfiguration('name');
     $file = $eqLogic->getConfiguration('img');
     $timestamp = microtime(true);
@@ -2561,7 +2604,7 @@ class frigate extends eqLogic
     $url = frigate::saveURL($uniqueId, null, $camera, 3, $file);
     $urlClip = "";
     // mise a jour des commandes
-    log::add(__CLASS__, 'debug', "| Mise à jour de la commande.");
+    log::add(__CLASS__, 'debug', "║ Mise à jour de la commande.");
     $eqLogic->getCmd(null, 'info_url_capture')->event("/var/www/html" . $url);
     $eqLogic->getCmd(null, 'info_label')->event("capture");
     $eqLogic->getCmd(null, 'info_score')->event(0);
@@ -2569,7 +2612,7 @@ class frigate extends eqLogic
     $eqLogic->getCmd(null, 'info_duree')->event(0);
 
     // Creation de l'evenement  dans la DB
-    log::add(__CLASS__, 'debug', "| Creéation d'un nouveau évènement Frigate pour l'event ID: " . $uniqueId);
+    log::add(__CLASS__, 'debug', "║ Creéation d'un nouveau évènement Frigate pour l'event ID: " . $uniqueId);
     $frigate = new frigate_events();
     $frigate->setCamera($camera);
     $frigate->setLasted($url);
@@ -2587,7 +2630,7 @@ class frigate extends eqLogic
     $frigate->setType("end");
     $frigate->setIsFavorite(0);
     $frigate->save();
-    log::add(__CLASS__, 'debug', "---------------------------------------------------");
+    log::add(__CLASS__, 'debug', "╚════════════════════════════════════════════════");
   }
 
   public static function createUniqueId($timestamp)
@@ -2729,8 +2772,8 @@ class frigate extends eqLogic
           $eventId = $value['after']['data']['detections'][0];
           $eventType = $value['type'];
           log::add(__CLASS__, 'info', ' => Traitement mqtt manual event <=');
-          log::add("frigate-zones", 'debug', 'Zone entrée : ' . $value['before']['zones'][0]);
-          log::add("frigate-zones", 'debug', 'Zone sortie : ' . $value['after']['zones'][0]);
+          //  log::add("frigate-zones", 'debug', 'Zone entrée : ' . $value['before']['zones'][0]);
+          //  log::add("frigate-zones", 'debug', 'Zone sortie : ' . $value['after']['zones'][0]);
 
           self::getEvent($eventId, $eventType);
           event::add('frigate::events', array('message' => 'mqtt_update_manual', 'type' => 'event'));
@@ -2954,7 +2997,7 @@ class frigate extends eqLogic
       // 0.15
       $curlResponse = $cResponse;
     }
-    
+
     $response = array(
       'status' => 'success',
       'message' => $curlResponse
@@ -3349,7 +3392,7 @@ class frigateCmd extends cmd
         $result = frigate::createEvent($camera, $eventParams['label'], $eventParams['video'], $eventParams['duration'], $eventParams['score']);
         $deamon_info = frigate::deamon_info();
         if ($deamon_info['launchable'] === 'nok') {
-          log::add('frigate', 'debug', "| action_make_api_event result = " . json_encode($result));
+          log::add('frigate', 'debug', "║ action_make_api_event result = " . json_encode($result));
           frigate::getEvent($result['event_id']);
         }
         break;
@@ -3358,13 +3401,13 @@ class frigateCmd extends cmd
         break;
       case 'action_http':
         // Gérer les variables user et password
-        log::add('frigate', 'info', "| 01 action_http $logicalId $link");
+        log::add('frigate', 'info', "║ 01 action_http $logicalId $link");
         $user = $frigate->getConfiguration("userName") ?? "";
         $password = $frigate->getConfiguration("password") ?? "";
         $link = str_replace("#user#", $user, $link);
         $link = str_replace("#password#", $password, $link);
         // Gérer les actions HTTP statiques
-        log::add('frigate', 'info', "| 02 action_http $logicalId $link");
+        log::add('frigate', 'info', "║ 02 action_http $logicalId $link");
         $response = self::getCurlcmd($link, $user, $password);
         if ($response !== false) {
           $frigate->getCmd(null, 'info_http')->event($response);
@@ -3403,13 +3446,13 @@ class frigateCmd extends cmd
     if (curl_errno($ch)) {
       log::add('frigate', 'error', "Erreur cURL: " . curl_error($ch));
     } else {
-      log::add('frigate', 'debug', "| Resultat de la commande HTTP : " . $response);
+      log::add('frigate', 'debug', "║ Resultat de la commande HTTP : " . $response);
     }
 
     // Récupérer les informations verbose
     rewind($verbose);
     $verboseLog = stream_get_contents($verbose);
-    log::add('frigate', 'debug', "| cURL verbose log: " . $verboseLog);
+    log::add('frigate', 'debug', "║ cURL verbose log: " . $verboseLog);
 
     fclose($verbose);
     curl_close($ch);