+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 19/02/2024 Beta 1.1.7
+- Gestion zone de sortie
+- Correction affichage fichier de configuration frigate > 0.15
+
 # 10/01/2024 Beta 1.1.6
 - Correction erreur cronDaily mySQL